% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{carga_datasets_ejemplo}
\alias{carga_datasets_ejemplo}
\title{Carga datasets de ejemplo}
\usage{
carga_datasets_ejemplo(name = "subvenciones")
}
\arguments{
\item{name}{Nombre del dataset de ejemplo a cargar (\code{"subvenciones"}
o \code{"convocatorias"}).}
}
\value{
Ver \code{\link[=carga_subvenciones]{carga_subvenciones()}}
}
\description{
Función para cargar datasets de ejemplo almacenados en el paquete.

\strong{Estos datos son un subconjunto reducido de la base de datos usados en
ejemplos y tests}. No deben ser usados para análisis.
}
\seealso{
\code{\link[=carga_subvenciones]{carga_subvenciones()}}, \code{\link[=carga_convocatorias]{carga_convocatorias()}}.
}
\keyword{internal}
