% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carga_datos.R
\name{carga_convocatorias}
\alias{carga_convocatorias}
\alias{carga_subvenciones}
\title{Carga ficheros del Volcado de la Base de Datos Nacional de Subvenciones
(BDNS).}
\usage{
carga_convocatorias(
  cache_dir = rappdirs::user_data_dir("subvencionesES"),
  update = FALSE,
  rebuild_db = FALSE
)

carga_subvenciones(
  cache_dir = rappdirs::user_data_dir("subvencionesES"),
  update = FALSE,
  rebuild_db = FALSE
)
}
\arguments{
\item{cache_dir}{Directorio donde se quiere almacenar los datos. Por
defecto \code{rappdirs::user_data_dir("subvencionesES")}. En este directorio
se almacena tambíén la copia procesada en formato .RData}

\item{update}{Variable lógica para descargar (\code{TRUE}) la versión más
actual o cargar la copia local. Por defecto carga copia local (\code{FALSE}).
Con la opción \code{update = TRUE} el paquete reescribe también el archivo
.RData asociado.}

\item{rebuild_db}{Vuelve a generar la base de datos local .RData a partir
de los archivos csv ya descargados.}
}
\value{
Un objeto en formato \code{tibble} con las cabeceras correctas.
}
\description{
Carga un objeto la información correspondiente. Se puede cargar una versión
previamente almacenada o bien descargar la última versión disponible.
\itemize{
\item \code{\link[=carga_convocatorias]{carga_convocatorias()}} carga información de convocatorias.
\item \code{\link[=carga_subvenciones]{carga_subvenciones()}} carga información de subvenciones.
}

La información se descarga en la carpeta especificada con el parámetro
\code{cache_dir}. Además, almacena una copia ya procesada de cada archivo en
formato .RData para reducir los tiempos de carga de datos en cada sesión
independiente.
}
\details{
Las funciones devuelven las columnas según los nombres explicados en el
issue \href{https://github.com/JaimeObregon/subvenciones/issues/11}{#11}.
\subsection{Resultados de \code{carga_convocatorias()}}{\tabular{ll}{
   \strong{nombres_columnas} \tab \strong{etiquetas_columnas} \cr
   IDConv \tab IDConv \cr
   id \tab Código BDNS \cr
   mrr \tab MRR \cr
   convocanteN1 \tab Administración \cr
   convocanteN2 \tab Departamento \cr
   convocanteN3 \tab Órgano \cr
   fechareg \tab Fecha de registro \cr
   titulo \tab Título de la convocatoria \cr
   bbreguladoras \tab URL de las BBRR \cr
   tituloleng \tab Título cooficial \cr
   verConcesiones \tab Ver concesiones \cr
   dummy1 \tab dummy1 \cr
   dummy2 \tab dummy2 \cr
}

}

\subsection{Resultados de \code{carga_subvenciones()}}{\tabular{ll}{
   \strong{nombres_columnas} \tab \strong{etiquetas_columnas} \cr
   ID \tab ID \cr
   IDConv \tab IDConvocatoria \cr
   convocanteN1 \tab Administración \cr
   convocanteN2 \tab Departamento \cr
   convocanteN3 \tab Órgano \cr
   convocatoria \tab Convocatoria \cr
   bbreguladoras \tab URL de las BBRR \cr
   programa \tab Aplicación presupuestaria \cr
   fechaconc \tab Fecha de concesión \cr
   beneficiario \tab Beneficiario \cr
   importe \tab Importe \cr
   instrumento \tab Instrumento \cr
   ayudaequiv \tab Ayuda Equivalente \cr
   detalles \tab Detalles \cr
   proyecto \tab proyecto \cr
   sancion \tab sancion \cr
   numcov \tab numcov \cr
}

}
}
\examples{
\dontrun{
conv <- carga_convocatorias()
sub <- carga_subvenciones()
}
}
\references{
Github repo: \url{https://github.com/JaimeObregon/subvenciones}.
}
\author{
Diego Hernangómez, Javier Álvarez-Liébana.
}
\concept{carga_datos}
