% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif_helpers.R
\name{extrae_cif}
\alias{extrae_cif}
\title{Extrae CIF del beneficiario}
\usage{
extrae_cif(x, col_benef = "beneficiario", append = TRUE)
}
\arguments{
\item{x}{Objeto inicial sobre el que se quiere ejecutar el split por CIF
(Obtenido o derivado de \code{\link[=carga_subvenciones]{carga_subvenciones()}}).}

\item{col_benef}{Nombre de la columna donde se almacena el valor de
\code{beneficiario}.}

\item{append}{Variable l칩gica para a침adir las nuevas columnas a \code{x} (\code{TRUE})
u obtener simplemente las columnas por separado (\code{FALSE}).}
}
\value{
Un \code{tibble} con las columnas \code{cif_beneficiario} y \code{nombre_beneficiario}.
Con la opci칩n \code{append = TRUE} se devuelve el objeto \code{x} inicial con ambas
columnas a침adidas.
}
\description{
Extrae CIF del beneficiario
}
\details{
Permite extraer el CIF y el nombre del beneficiario (sin el CIF) de
un registro en concreto
}
\examples{

# Carga dataset ejemplo
sub <- carga_datasets_ejemplo("subvenciones")

# Para carga completa usar:
# sub <- carga_subvenciones()

# Extrae CIF

library(dplyr)

extrae_cif(sub) \%>\%
  select(IDConv, fechaconc, importe, cif_beneficiario, nombre_beneficiario)
}
\seealso{
\code{\link[=carga_subvenciones]{carga_subvenciones()}}

Otras funciones de ayuda para CIFs:
\code{\link{valida_cif}()}
}
\concept{cif_helpers}
