% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subvencionesES-package.R
\docType{package}
\name{subvencionesES-package}
\alias{subvencionesES}
\alias{subvencionesES-package}
\title{subvencionesES: Descarga la información de la Base de Base de Datos Nacional de Subvenciones ('BDNS')}
\description{
Aportación desde la comunidad de 'R' para la extracción, limpieza y visualización de los datos del volcado de la Base de Datos Nacional de Subvenciones ('BDNS'). Github del proyecto global y comunitario puesto en marcha por Jaime Gómez-Obregón <https://github.com/JaimeObregon/subvenciones>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dadosdelaplace/Rsubvenciones_spain/tree/rpak}
  \item Report bugs at \url{https://github.com/dadosdelaplace/Rsubvenciones_spain/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{afakemail@donotuse.com}

Authors:
\itemize{
  \item Javier Álvarez Liébana
}

}
\keyword{internal}
