% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{valida_cif}
\alias{valida_cif}
\title{Valida CIFs}
\usage{
valida_cif(cif)
}
\arguments{
\item{cif}{Vector de Códigos CIF a validar.}
}
\value{
Un vector lógico \code{TRUE/FALSE}
}
\description{
Valida CIFs españoles.
}
\examples{

# Carga dataset ejemplo
sub <- carga_datasets_ejemplo("subvenciones")

# Para carga completa usar:
# sub <- carga_subvenciones()


library(dplyr)

nifs <- sub \%>\%
  extrae_cif(append = FALSE) \%>\%
  mutate(valid = valida_cif(cif_beneficiario))


nifs

}
\references{
\url{https://es.wikipedia.org/wiki/Código_de_identificación_fiscal}
}
\seealso{
Other helpers: 
\code{\link{extrae_cif}()}
}
\concept{helpers}
